+++ src/main/java/de/challenge/challenges/randomizer/CraftingRandomizerChallenge.java
@@ -1,6 +1,7 @@
 package de.challenge.challenges.randomizer;
 
 import de.challenge.Challenge;
+import de.challenge.ChallengeCategory;
 import de.challenge.ChallengePlugin;
 import de.challenge.util.RandomizerMap;
 import org.bukkit.Bukkit;
@@ -36,6 +37,9 @@ public class CraftingRandomizerChallenge extends Challenge {
 
     @Override
     public String getDescription() { return "Crafting results are randomized"; }
+
+    @Override
+    public ChallengeCategory getCategory() { return ChallengeCategory.RANDOMIZER; }
 
     @Override
     protected void onEnable() {
@@ -98,7 +102,10 @@ public class CraftingRandomizerChallenge extends Challenge {
     @Override
     protected void onDisable() {
         if (craftMap != null) {
-            craftMap.delete();
+            if (!preservingState) {
+                craftMap.delete();
+            }
+            craftMap = null;
         }
     }
 }
